refactor: users router, organizations router
