+++ packages/back-end/src/routers/organizations/organizations.controller.ts
@@ -1,24 +1,21 @@
-import { Request, Response } from "express";
-import { AuthRequest } from "../types/AuthRequest";
+import { Response } from "express";
+import { AuthRequest } from "../../types/AuthRequest";
 import {
   acceptInvite,
+  getInviteUrl,
+  getOrgFromReq,
+  importConfig,
   inviteUser,
+  isEnterpriseSSO,
   removeMember,
   revokeInvite,
-  getInviteUrl,
-  importConfig,
-  getOrgFromReq,
-  addMemberFromSSOConnection,
-  isEnterpriseSSO,
-  validateLoginMethod,
-} from "../services/organizations";
+} from "../../services/organizations";
 import {
-  getSourceIntegrationObject,
   getNonSensitiveParams,
-} from "../services/datasource";
-import { createUser, getUsersByIds, updatePassword } from "../services/users";
-import { getAllTags } from "../models/TagModel";
-import { UserModel } from "../models/UserModel";
+  getSourceIntegrationObject,
+} from "../../services/datasource";
+import { getUsersByIds, updatePassword } from "../../services/users";
+import { getAllTags } from "../../models/TagModel";
 import {
   ExpandedMember,
   Invite,
@@ -26,42 +23,37 @@ import {
   NamespaceUsage,
   OrganizationInterface,
   OrganizationSettings,
-} from "../../types/organization";
+} from "../../../types/organization";
 import {
-  getWatchedAudits,
+  auditDetailsUpdate,
   findByEntity,
   findByEntityParent,
-  auditDetailsUpdate,
-} from "../services/audit";
-import { WatchModel } from "../models/WatchModel";
-import { ExperimentModel } from "../models/ExperimentModel";
-import { getExperimentById, ensureWatching } from "../services/experiments";
-import { getFeature, getAllFeatures } from "../models/FeatureModel";
-import { SegmentModel } from "../models/SegmentModel";
-import { findDimensionsByOrganization } from "../models/DimensionModel";
-import { IS_CLOUD } from "../util/secrets";
-import { sendInviteEmail, sendNewOrgEmail } from "../services/email";
-import { getDataSourcesByOrganization } from "../models/DataSourceModel";
-import { getAllGroups } from "../services/group";
-import { getAllSavedGroups } from "../models/SavedGroupModel";
-import { uploadFile } from "../services/files";
-import { getMetricsByOrganization } from "../models/MetricModel";
-import { WebhookModel } from "../models/WebhookModel";
-import { createWebhook } from "../services/webhooks";
+  getWatchedAudits,
+} from "../../services/audit";
+import { ExperimentModel } from "../../models/ExperimentModel";
+import { getAllFeatures } from "../../models/FeatureModel";
+import { SegmentModel } from "../../models/SegmentModel";
+import { findDimensionsByOrganization } from "../../models/DimensionModel";
+import { IS_CLOUD } from "../../util/secrets";
+import { sendInviteEmail, sendNewOrgEmail } from "../../services/email";
+import { getDataSourcesByOrganization } from "../../models/DataSourceModel";
+import { getAllGroups } from "../../services/group";
+import { getAllSavedGroups } from "../../models/SavedGroupModel";
+import { getMetricsByOrganization } from "../../models/MetricModel";
+import { WebhookModel } from "../../models/WebhookModel";
+import { createWebhook } from "../../services/webhooks";
 import {
   createOrganization,
-  findOrganizationsByMemberId,
   hasOrganization,
   updateOrganization,
-} from "../models/OrganizationModel";
-import { findAllProjectsByOrganization } from "../models/ProjectModel";
-import { ConfigFile } from "../init/config";
-import { WebhookInterface } from "../../types/webhook";
-import { ExperimentRule, NamespaceValue } from "../../types/feature";
-import { usingOpenId } from "../services/auth";
+} from "../../models/OrganizationModel";
+import { findAllProjectsByOrganization } from "../../models/ProjectModel";
+import { ConfigFile } from "../../init/config";
+import { WebhookInterface } from "../../../types/webhook";
+import { ExperimentRule, NamespaceValue } from "../../../types/feature";
+import { usingOpenId } from "../../services/auth";
 import { cloneDeep } from "lodash";
-import { getLicense } from "../init/license";
-import { getSSOConnectionSummary } from "../models/SSOConnectionModel";
+import { getSSOConnectionSummary } from "../../models/SSOConnectionModel";
 import {
   createApiKey,
   deleteApiKeyById,
@@ -70,70 +62,12 @@ import {
   getApiKeyByIdOrKey,
   getFirstPublishableApiKey,
   getUnredactedSecretKey,
-} from "../models/ApiKeyModel";
+} from "../../models/ApiKeyModel";
 import {
   accountFeatures,
   getAccountPlan,
   getRoles,
-} from "../util/organization.util";
-
-export async function getUser(req: AuthRequest, res: Response) {
-  // If using SSO, auto-create users in Mongo who we don't recognize yet
-  if (!req.userId && usingOpenId()) {
-    const user = await createUser(req.name || "", req.email, "", req.verified);
-    req.userId = user.id;
-  }
-
-  if (!req.userId) {
-    throw new Error("Must be logged in");
-  }
-
-  const userId = req.userId;
-
-  // List of all organizations the user belongs to
-  const orgs = await findOrganizationsByMemberId(userId);
-
-  // If the user is not in an organization yet and is using SSO
-  // Check to see if they should be auto-added to one based on their email domain
-  if (!orgs.length) {
-    const autoOrg = await addMemberFromSSOConnection(req);
-    if (autoOrg) {
-      orgs.push(autoOrg);
-    }
-  }
-
-  // Filter out orgs that the user can't log in to
-  let lastError = "";
-  const validOrgs = orgs.filter((org) => {
-    try {
-      validateLoginMethod(org, req);
-      return true;
-    } catch (e) {
-      lastError = e;
-      return false;
-    }
-  });
-
-  // If all of a user's orgs were filtered out, throw an error
-  if (orgs.length && !validOrgs.length) {
-    throw new Error(lastError || "Must login with SSO");
-  }
-
-  return res.status(200).json({
-    status: 200,
-    userId: userId,
-    userName: req.name,
-    email: req.email,
-    admin: !!req.admin,
-    license: !IS_CLOUD && getLicense(),
-    organizations: validOrgs.map((org) => {
-      return {
-        id: org.id,
-        name: org.name,
-      };
-    }),
-  });
-}
+} from "../../util/organization.util";
 
 export async function getDefinitions(req: AuthRequest, res: Response) {
   const { org } = getOrgFromReq(req);
@@ -249,92 +183,6 @@ export async function getActivityFeed(req: AuthRequest, res: Response) {
   }
 }
 
-export async function postWatchItem(
-  req: AuthRequest<null, { type: string; id: string }>,
-  res: Response
-) {
-  const { org, userId } = getOrgFromReq(req);
-  const { type, id } = req.params;
-  let item;
-
-  if (type === "feature") {
-    item = await getFeature(org.id, id);
-  } else if (type === "experiment") {
-    item = await getExperimentById(id);
-    if (item && item.organization !== org.id) {
-      res.status(403).json({
-        status: 403,
-        message: "You do not have access to this experiment",
-      });
-      return;
-    }
-  }
-  if (!item) {
-    throw new Error(`Could not find ${item}`);
-  }
-  if (type == "feature") {
-    await ensureWatching(userId, org.id, id, "features");
-  } else {
-    await ensureWatching(userId, org.id, id, "experiments");
-  }
-
-  return res.status(200).json({
-    status: 200,
-  });
-}
-
-export async function postUnwatchItem(
-  req: AuthRequest<null, { type: string; id: string }>,
-  res: Response
-) {
-  const { org, userId } = getOrgFromReq(req);
-  const { type, id } = req.params;
-  const pluralType = type + "s";
-
-  try {
-    await WatchModel.updateOne(
-      {
-        userId: userId,
-        organization: org.id,
-      },
-      {
-        $pull: {
-          [pluralType]: id,
-        },
-      }
-    );
-
-    return res.status(200).json({
-      status: 200,
-    });
-  } catch (e) {
-    res.status(400).json({
-      status: 400,
-      message: e.message,
-    });
-  }
-}
-
-export async function getWatchedItems(req: AuthRequest, res: Response) {
-  const { org, userId } = getOrgFromReq(req);
-  try {
-    const watch = await WatchModel.findOne({
-      userId: userId,
-      organization: org.id,
-    });
-    res.status(200).json({
-      status: 200,
-      experiments: watch?.experiments || [],
-      features: watch?.features || [],
-    });
-  } catch (e) {
-    res.status(400).json({
-      status: 400,
-      message: e.message,
-    });
-  }
-}
-
 export async function getHistory(
   req: AuthRequest<null, { type: string; id: string }>,
   res: Response
@@ -366,35 +214,6 @@ export async function getHistory(
     status: 200,
     events: merged,
   });
-}
-
-export async function putUserName(
-  req: AuthRequest<{ name: string }>,
-  res: Response
-) {
-  const { name } = req.body;
-  const { userId } = getOrgFromReq(req);
-
-  try {
-    await UserModel.updateOne(
-      {
-        id: userId,
-      },
-      {
-        $set: {
-          name,
-        },
-      }
-    );
-    res.status(200).json({
-      status: 200,
-    });
-  } catch (e) {
-    res.status(400).json({
-      status: 400,
-      message: e.message || "An error occurred",
-    });
-  }
 }
 
 export async function putMemberRole(
@@ -1264,15 +1083,6 @@ export async function postImportConfig(
   }
 
   await importConfig(config, org);
-
-  res.status(200).json({
-    status: 200,
-  });
-}
-
-export async function putUpload(req: Request, res: Response) {
-  const { signature, path } = req.query as { signature: string; path: string };
-  await uploadFile(path, signature, req.body);
 
   res.status(200).json({
     status: 200,
